Desfazer alteração relativa ao problema de encoding no ano/turma do 1ºciclo

+++ admin/includes/ano_escolar.php
@@ -1,7 +1,7 @@
 <?php
 
 	if($_GET['save'] == 1){
-		$ano_escolar = utf8_decode($_POST[ano_escolar]);
+		$ano_escolar = $_POST[ano_escolar];
 		$ref = $_POST[ref_ano];
 		$res = mysql_query("UPDATE `dbo_tab_anos_escolares` SET `ANO_ESCOLAR`='$ano_escolar',`REFERENCIA`='$ref' WHERE ID_ANO_ESCOLAR = $_GET[id]");
 
@@ -9,17 +9,17 @@
 	}
 
 	if($_GET['delete']==1){
-		//$search = mysql_query("SELECT * from dbo_tab_primeiro_ciclo where ID_ANO_ESCOLAR = $_GET[id] limit 1");
-		//$num_rows = mysql_num_rows($search);
+		$search = mysql_query("SELECT * from dbo_tab_primeiro_ciclo where ID_ANO_ESCOLAR = $_GET[id] limit 1");
+		$num_rows = mysql_num_rows($search);
 
-		//if ($num_rows == 0) {
+		if ($num_rows == 0) {
 			$res = mysql_query("DELETE FROM `dbo_tab_anos_escolares` WHERE ID_ANO_ESCOLAR = $_GET[id]");
-		/*}else if ($num_rows == 1 && !empty($_GET['id'])){
+		}else if ($num_rows == 1 && !empty($_GET['id'])){
 			echo '<div class="alert alert-warning">
 				<a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
 			  <a data-dismiss="alert"><strong>Atenção!</strong> Não é possível apagar um ano escolar que já tenha registos efectuados!</a>
 			</div>';
-		}*/
+		}
 	}
 ?>
 <div class="row">
@@ -41,7 +41,7 @@
 					while($row = mysql_fetch_object($res)){ ?>
 						<tr>
 							<td> <?php echo $row->REFERENCIA; ?> </td>
-							<td> <?php echo utf8_encode($row->ANO_ESCOLAR); ?> </td>
+							<td> <?php echo $row->ANO_ESCOLAR; ?> </td>
 							<td style='text-align: center;'>
 								<a href='index.php?mod=edit_ano_esc&id=<?php echo $row->ID_ANO_ESCOLAR; ?>' class='btn btn-xs btn-warning'>
 									<span class='glyphicon glyphicon-pencil'></span>
